requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiohttp-swagger==1.0.16
aiolimiter==1.1.0
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
asyncpg==0.29.0
attrs==23.2.0
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.5
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Jinja2==3.1.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
magic-filter==1.0.12
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
mypy==1.9.0
mypy-extensions==1.0.0
OSlash==0.6.3
packaging==23.2
psutil==6.0.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyOpenSSL==24.0.0
pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-telegram-bot==20.8
pytz==2024.1
PyYAML==6.0.2
redis==5.1.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
service-identity==24.1.0
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
telebot==0.0.5
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.27.1
wheel==0.43.0
yarl==1.9.4
zope.interface==6.2
pydub
pywhatkit
SpeechRecognition
pyaudio